third_party/blink/web_tests/svg/filters/feComponentTransfer-style-crash-expected.txt

ALERT: TypeError: Failed to set the 'baseVal' property on 'SVGAnimatedEnumeration': The enumeration value provided (34463) is larger than the largest allowed value (5).
PASS
